fix: new file for email modal
