+++ src/Components/DnDEditor/DnDEditor.css
@@ -29,6 +29,7 @@
 }
 
 .send-btn-group {
+  margin: 15px 0 0 0;
   text-align: center;
 }
